% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{tr_dependency_core}
\alias{tr_dependency_core}
\title{Dependência Shiny com os assets do núcleo (CSS do core e do xyflow, \code{runtime.js}, \code{editor.js}).}
\usage{
tr_dependency_core(ver = as.character(utils::packageVersion("trama")))
}
\description{
Dependência Shiny com os assets do núcleo (CSS do core e do xyflow, \code{runtime.js}, \code{editor.js}).
}
