% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{tr_dependency_importmap}
\alias{tr_dependency_importmap}
\title{Dependência Shiny que injeta o importmap e a captura de erro de boot, antes de qualquer módulo.}
\usage{
tr_dependency_importmap(base)
}
\description{
Dependência Shiny que injeta o importmap e a captura de erro de boot, antes de qualquer módulo.
}
