% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-io.R
\name{tr_doc_parse}
\alias{tr_doc_parse}
\title{Lê um documento a partir de texto JSON, validando o formato e reconstituindo vetores achatados pelo parse.}
\usage{
tr_doc_parse(txt)
}
\description{
Lê um documento a partir de texto JSON, validando o formato e reconstituindo vetores achatados pelo parse.
}
