% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{tr_doc_terminals}
\alias{tr_doc_terminals}
\title{Nós sem aresta de saída — os alvos default de execução.}
\usage{
tr_doc_terminals(doc)
}
\description{
Nós sem aresta de saída — os alvos default de execução.
}
