% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{tr_get_node}
\alias{tr_get_node}
\title{Busca um \code{tr_node} pelo id qualificado (\code{colecao/nome}); erro alto se não registrado.}
\usage{
tr_get_node(id, registry = .tr_default_registry)
}
\description{
Busca um \code{tr_node} pelo id qualificado (\code{colecao/nome}); erro alto se não registrado.
}
