% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{tr_get_type}
\alias{tr_get_type}
\title{Busca um \code{tr_type} pelo id qualificado; erro alto se não registrado.}
\usage{
tr_get_type(id, registry = .tr_default_registry)
}
\description{
Busca um \code{tr_type} pelo id qualificado; erro alto se não registrado.
}
