% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{tr_handle_failed}
\alias{tr_handle_failed}
\title{Um handle representa erro (não valor) quando carrega \verb{$error}.}
\usage{
tr_handle_failed(handle)
}
\description{
Um handle representa erro (não valor) quando carrega \verb{$error}.
}
