% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{tr_is_template}
\alias{tr_is_template}
\title{O texto é um template do trama? Nunca falha: JSON inválido é só "não".}
\usage{
tr_is_template(txt)
}
\description{
O texto é um template do trama? Nunca falha: JSON inválido é só "não".
}
