% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{tr_link}
\alias{tr_link}
\title{Liga \code{from} a \code{to}. Porta omitida = primeira saída / primeira entrada.}
\usage{
tr_link(flow, from, to, index = NULL)
}
\description{
Liga \code{from} a \code{to}. Porta omitida = primeira saída / primeira entrada.
}
