% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{tr_param_num}
\alias{tr_param_num}
\title{Param numérico (ponto flutuante).}
\usage{
tr_param_num(
  default,
  min = NA,
  max = NA,
  step = NULL,
  label = NULL,
  unit = NA_character_
)
}
\description{
Param numérico (ponto flutuante).
}
