% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{tr_runs}
\alias{tr_runs}
\title{Lê todos os runs de um projeto como data.frame (um evento por linha).}
\usage{
tr_runs(project)
}
\description{
Lê todos os runs de um projeto como data.frame (um evento por linha).
}
