% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{tr_set}
\alias{tr_set}
\title{Aplica \code{set_param} para cada \code{...} nomeado, um por um.}
\usage{
tr_set(flow, id, ...)
}
\description{
Aplica \code{set_param} para cada \code{...} nomeado, um por um.
}
