% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{tr_template_save}
\alias{tr_template_save}
\title{Grava o template em \code{dir}; nome de arquivo derivado do nome do template.}
\usage{
tr_template_save(tpl, dir, overwrite = FALSE)
}
\description{
Grava o template em \code{dir}; nome de arquivo derivado do nome do template.
}
