% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{tr_ui}
\alias{tr_ui}
\title{Monta a UI Shiny do editor: importmap, dependências e o \verb{<div>} onde o React monta.}
\usage{
tr_ui(project)
}
\description{
Monta a UI Shiny do editor: importmap, dependências e o \verb{<div>} onde o React monta.
}
