% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{tr_flow_doc}
\alias{tr_flow_doc}
\title{O \code{tr_doc} por trás de um \code{tr_flow} — a saída pronta pra \code{tr_run()}, \code{tr_doc_write()} etc.}
\usage{
tr_flow_doc(flow)
}
\description{
O \code{tr_doc} por trás de um \code{tr_flow} — a saída pronta pra \code{tr_run()}, \code{tr_doc_write()} etc.
}
