% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{tr_project_save}
\alias{tr_project_save}
\title{Grava o flow \code{name} do projeto em \verb{flows/<name>.json}.}
\usage{
tr_project_save(project, doc, name = "main")
}
\description{
Grava o flow \code{name} do projeto em \verb{flows/<name>.json}.
}
